feat: added documentation for theme toggle animations
