--- apps/zenithui-docs/src/components/theme-animator/pixelate.tsx
@@ -1,19 +1,0 @@
-import { useThemeMode } from "@zenithui/utils"
-
-export default function RandomBlocksThemeAnimator() {
-  const { ref, toggleSwitchTheme, isDarkMode } = useThemeMode({
-    animationType: "PIXELATE",
-    duration: 750,
-    blurAmount: 2,
-  })
-
-  return (
-    <button
-      type="button"
-      ref={ref}
-      onClick={toggleSwitchTheme}
-    >
-      {isDarkMode ? "🌙" : "☀️"}
-    </button>
-  )
-}
