--- apps/zenithui-docs/src/components/theme-animator/spiral.tsx
@@ -1,19 +1,0 @@
-import { useThemeMode } from "@zenithui/utils"
-
-export default function SpiralThemeAnimator() {
-  const { ref, toggleSwitchTheme, isDarkMode } = useThemeMode({
-    animationType: "SPIRAL",
-    duration: 750,
-    blurAmount: 2,
-  })
-
-  return (
-    <button
-      type="button"
-      ref={ref}
-      onClick={toggleSwitchTheme}
-    >
-      {isDarkMode ? "🌙" : "☀️"}
-    </button>
-  )
-}
