--- apps/zenithui-docs/src/content/theme-animator/theme-animotor.mdx
@@ -1,67 +1,0 @@
-import CircleThemeAnimator from "@/components/theme-animator/circle"
-import CircleBlurThemeAnimator from "@/components/theme-animator/circle-blur"
-import SwipeThemeAnimator from "@/components/theme-animator/swipe"
-import DiamondThemeAnimator from "@/components/theme-animator/diamond"
-import RippleThemeAnimator from "@/components/theme-animator/ripple"
-import ShutterThemeAnimator from "@/components/theme-animator/shutter"
-import SplitThemeAnimator from "@/components/theme-animator/split"
-import SpiralThemeAnimator from "@/components/theme-animator/spiral"
-import PixelateThemeAnimator from "@/components/theme-animator/pixelate"
-import RandomBlocksThemeAnimator from "@/components/theme-animator/random-blocks"
-import CornersThemeAnimator from "@/components/theme-animator/corners"
-import WarpThemeAnimator from "@/components/theme-animator/warp"
-import GlitchThemeAnimator from "@/components/theme-animator/glitch"
-
-# Animations
-
-## Circle
-
-<CircleThemeAnimator />
-
-## Circle Blur
-
-<CircleBlurThemeAnimator />
-
-## Swipe
-
-<SwipeThemeAnimator />
-
-## Diamond
-
-<DiamondThemeAnimator />
-
-## Ripple
-
-<RippleThemeAnimator />
-
-## Shutter
-
-<ShutterThemeAnimator />
-
-## Split
-
-<SplitThemeAnimator />
-
-## Spiral
-
-<SpiralThemeAnimator />
-
-## Pixelate
-
-<PixelateThemeAnimator />
-
-## Random Blocks
-
-<RandomBlocksThemeAnimator />
-
-## Corners
-
-<CornersThemeAnimator />
-
-## Warp
-
-<WarpThemeAnimator />
-
-## Glitch
-
-<GlitchThemeAnimator />
