+++ apps/zenithui-docs/src/content/theme-toggle/_meta.js
