+++ apps/zenithui-docs/src/content/theme-toggle/animations.mdx
@@ -1,0 +1,124 @@
+import CircleThemeAnimator from "@/components/theme-toggle/circle"
+import CircleBlurThemeAnimator from "@/components/theme-toggle/circle-blur"
+import SwipeThemeAnimator from "@/components/theme-toggle/swipe"
+import DiamondThemeAnimator from "@/components/theme-toggle/diamond"
+import RippleThemeAnimator from "@/components/theme-toggle/ripple"
+import ShutterThemeAnimator from "@/components/theme-toggle/shutter"
+import SplitThemeAnimator from "@/components/theme-toggle/split"
+import SpiralThemeAnimator from "@/components/theme-toggle/spiral"
+import PixelateThemeAnimator from "@/components/theme-toggle/pixelate"
+import RandomBlocksThemeAnimator from "@/components/theme-toggle/random-blocks"
+import CornersThemeAnimator from "@/components/theme-toggle/corners"
+import WarpThemeAnimator from "@/components/theme-toggle/warp"
+import GlitchThemeAnimator from "@/components/theme-toggle/glitch"
+
+# Available Animations
+
+Below are all supported theme transition animations with corresponding components.
+
+## 1. Circle
+
+A smooth circular expansion effect from the click/toggle point.
+
+<CircleThemeAnimator />
+
+## 2. Circle Blur
+
+Adds a blur overlay expanding in a circular motion.
+
+<CircleBlurThemeAnimator />
+
+## 3. Swipe
+
+Slides the new theme from one side of the screen to the other.
+
+<SwipeThemeAnimator />
+
+## 4. Diamond
+
+Reveals the new theme through a diamond-shaped transition.
+
+<DiamondThemeAnimator />
+
+## 5. Ripple
+
+Emulates a water ripple effect during the theme change.
+
+<RippleThemeAnimator />
+
+## 6. Shutter
+
+Imitates camera shutters closing and opening to switch themes.
+
+<ShutterThemeAnimator />
+
+## 7. Split
+
+Splits the screen into two halves that slide apart and reveal the new theme.
+
+<SplitThemeAnimator />
+
+## 8. Spiral
+
+Swirls the transition in a spiral motion for a dynamic effect.
+
+<SpiralThemeAnimator />
+
+## 9. Pixelate
+
+Creates a pixelation effect as the theme changes.
+
+<PixelateThemeAnimator />
+
+## 10. Random Blocks
+
+Displays randomized blocks revealing the new theme progressively.
+
+<RandomBlocksThemeAnimator />
+
+## 11. Corners
+
+Animates the transition starting from all four screen corners.
+
+<CornersThemeAnimator />
+
+## 12. Warp
+
+Applies a distortion-like warp animation across the screen.
+
+<WarpThemeAnimator />
+
+## 13. Glitch
+
+Simulates a digital glitch effect while switching themes.
+
+<GlitchThemeAnimator />
+
+---
+
+### Best Practices
+
+- Keep animation duration under 800ms for responsiveness.
+- Use subtle effects (like Circle or Ripple) for production apps.
+- Reserve intense animations (like Glitch or Pixelate) for creative or portfolio sites.
+- Always ensure accessibility — avoid rapid flashing or over-bright transitions.
+
+---
+
+### Summary Table
+
+| Animation     | Description                        | Recommended Use               |
+| ------------- | ---------------------------------- | ----------------------------- |
+| Circle        | Expanding circle from center/click | Balanced, minimal apps        |
+| Circle Blur   | Blurred circle expansion           | Modern, aesthetic UIs         |
+| Swipe         | Horizontal slide                   | Dashboard transitions         |
+| Diamond       | Diamond shape reveal               | Geometric/visual apps         |
+| Ripple        | Water-like ripple                  | Clean and elegant transitions |
+| Shutter       | Camera shutter motion              | Cinematic UI effects          |
+| Split         | Split reveal from center           | Dual-theme designs            |
+| Spiral        | Rotating spiral                    | Experimental/creative designs |
+| Pixelate      | Pixel block reveal                 | Retro or techy themes         |
+| Random Blocks | Random square reveal               | Dynamic UIs                   |
+| Corners       | From all corners                   | Symmetric transitions         |
+| Warp          | Distorted warp                     | Sci-fi or 3D UIs              |
+| Glitch        | Digital glitch                     | Futuristic/cyberpunk UIs      |
