+++ apps/zenithui-docs/src/content/theme-toggle/getting-started.mdx
